refactor: move provider helpers to dal/env.py and remove factory/ directory
